requirements.txt
customtkinter
pystray
Pillow
ytmusicapi
faster-whisper
keyboard
sounddevice
soundfile
numpy
